+++ thoughtstream-frontend/src/components/DiaryList.jsx
@@ -1,20 +1,52 @@
-// src/components/DiaryEntryCard.jsx
-import React from 'react';
+import React, { useEffect, useState } from "react";
+import DiaryEntryCard from "./DiaryEntryCard";
 
-function DiaryEntryCard({ entry }) {
-  const { title, content, createdAt, weather } = entry;
+function DiaryList() {
+  const [entries, setEntries] = useState([]);
+  const [error, setError] = useState(null);
 
-  // Format the creation timestamp
-  const formattedDate = new Date(createdAt).toLocaleString();
+  useEffect(() => {
+    const fetchEntries = async () => {
+      const token = localStorage.getItem("jwt");
+      if (!token) {
+        setError("Please log in to view your entries.");
+        return;
+      }
+
+      try {
+        const response = await fetch(`${import.meta.env.VITE_API_BASE_URL}/api/diary`, {
+          method: "GET",
+          headers: {
+            Authorization: `Bearer ${token}`,
+          },
+        });
+
+        if (!response.ok) {
+          throw new Error("Failed to fetch entries.");
+        }
+
+        const data = await response.json();
+        setEntries(data.entries || []);
+      } catch (err) {
+        setError(err.message);
+      }
+    };
+
+    fetchEntries();
+  }, []);
+
+  if (error) return <p>{error}</p>;
+  if (!entries.length) return <p>No diary entries found.</p>;
 
   return (
-    <div className="diary-entry-card">
-      <h3>{title}</h3>
-      <p>{content}</p>
-      {weather && <p><strong>Weather:</strong> {weather}</p>}
-      <p><small>Created on: {formattedDate}</small></p>
-    </div>
+    <ul className="diaryList">
+      {entries.map((entry) => (
+        <li key={entry._id || entry.id}>
+          <DiaryEntryCard entry={entry} />
+        </li>
+      ))}
+    </ul>
   );
 }
 
-export default DiaryEntryCard;
+export default DiaryList;
